+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+# 0.2.0 (2017-09-21)
+
+Renamed `Identity` to `LoginGov` because `Identity` because of a name collision with the `identities` table ActiveRecord model in `identity-idp`
+
+- Added `LoginGov::Hostdata.domain`
+- Added `LoginGov::Hostdata.env`
+
 # 0.1.0 (2017-09-21)
 
 - Added `Identity::Hostdata.domain`